refactor(memory): split rpc and completion modules
